requirements.txt
gradio
ollama
Pillow
sqlmodel